# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: ChIAtraton HTTP API
  version: 1.0.0
  summary: Contractul API pentru workspace-ul de verificare AI.
  description: |
    API contract-first pentru verificarea unui raport periodic selectat.
    ChIAtraton nu înlocuiește MyADR/MySMIS și nu execută acțiuni oficiale.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Originea mediului în care este publicat API-ul.
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/reports/{reportId}/analysis-jobs:
    post:
      tags: [Analysis]
      operationId: createReportAnalysisJob
      summary: Pornește analiza asincronă a raportului.
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisJobCreate'
      responses:
        '202':
          description: Analiza a fost acceptată în coadă.
          headers:
            Location:
              $ref: '#/components/headers/Location'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/AIUnavailable'
  /api/v1/analysis-jobs/{jobId}:
    get:
      tags: [Analysis]
      operationId: getAnalysisJob
      summary: Citește starea unui job de analiză.
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Starea jobului.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/criterion-extraction-jobs/{jobId}/proposals:
    get:
      tags: [Criterion extraction]
      operationId: listCriterionExtractionProposals
      summary: Listează propunerile produse de un job de extracție finalizat.
      description: |
        Returnează 409 invalid_analysis_job_kind pentru alt tip de job și
        409 analysis_job_not_succeeded dacă extracția nu s-a încheiat cu succes.
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Pagină de propuneri, inclusiv review-ul lor când există.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCriterionProposals'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/criterion-extraction-jobs/{jobId}/proposal-reviews:
    post:
      tags: [Criterion extraction]
      operationId: createCriterionProposalReviews
      summary: Salvează review-uri umane pentru propunerile jobului.
      description: |
        Batch-ul este atomic. accept și correct creează câte un Criterion nou;
        reject păstrează propunerea și review-ul, fără să creeze Criterion.
        Conflictele folosesc proposal_already_reviewed, stale_proposal_revision,
        criterion_code_conflict sau idempotency_conflict.
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriterionProposalReviewBatch'
      responses:
        '201':
          description: Review-urile au fost salvate și sunt auditabile.
          headers:
            Location:
              $ref: '#/components/headers/Location'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CriterionProposalReviewBatchResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Tokenul identifică utilizatorul tehnic pentru audit.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Cheie opacă pentru replay sigur al operației POST.
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Cursor opac returnat de pagina precedentă.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
...
  headers:
    Location:
      description: URI-ul resursei create.
      schema:
        type: string
        format: uri-reference
    IdempotencyReplayed:
      description: Este true numai când răspunsul este un replay idempotent.
      required: false
      schema:
        type: boolean
    RequestId:
      description: Identificator opac pentru corelarea tehnică a cererii.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    Unauthorized:
      description: Token absent sau invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resursa nu există sau nu este vizibilă utilizatorului.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Conflict de stare, idempotency sau unicitate.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    ValidationError:
      description: Request invalid semantic sau structural.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InternalError:
      description: Eroare internă fără date sensibile.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    AIUnavailable:
      description: Serviciul AI nu este disponibil pentru pornirea analizei.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    SourceAnchor:
      type: object
      additionalProperties: false
      required: [documentId, pageNumber, passage]
      properties:
        documentId:
          type: string
          format: uuid
        pageNumber:
          type: integer
          minimum: 1
        passage:
          type: string
          minLength: 1
          maxLength: 8000
...
    Criterion:
      type: object
      additionalProperties: false
      required:
        - id
        - projectId
        - code
        - description
        - deadline
        - sourceAnchors
        - version
        - active
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        code:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 4000
        deadline:
          type: [string, 'null']
          format: date
        sourceAnchors:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/SourceAnchor'
        version:
          type: integer
          minimum: 1
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AnalysisJobCreate:
      type: object
      additionalProperties: false
      properties:
        projectDocumentIds:
          type: array
          uniqueItems: true
          maxItems: 100
          default: []
          items:
            type: string
            format: uuid

        previousReportIds:
          type: array
          uniqueItems: true
          maxItems: 100
          default: []
          items:
            type: string
            format: uuid
    AnalysisJobKind:
      type: string
      enum: [analyze_report, extract_criteria]
    AnalysisJobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    AnalysisJobError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 500
    AnalysisJob:
      type: object
      additionalProperties: false
      required:
        - id
        - projectId
        - kind
        - reportId
        - status
        - documentIds
        - projectDocumentIds
        - previousReportIds
        - criteriaSnapshotVersion
        - proposalCount
        - createdAt
        - startedAt
        - completedAt
        - error
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/AnalysisJobKind'
        reportId:
          type: [string, 'null']
          format: uuid
        status:
          $ref: '#/components/schemas/AnalysisJobStatus'
        documentIds:
          type: array
          uniqueItems: true
          maxItems: 100
          items:
            type: string
            format: uuid
        projectDocumentIds:
          type: array
          uniqueItems: true
          maxItems: 100
          items:
            type: string
            format: uuid
        previousReportIds:
          type: array
          uniqueItems: true
          maxItems: 100
          items:
            type: string
            format: uuid
        criteriaSnapshotVersion:
          type: [integer, 'null']
          minimum: 1
        proposalCount:
          type: [integer, 'null']
          minimum: 0
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        error:
          oneOf:
            - $ref: '#/components/schemas/AnalysisJobError'
            - type: 'null'
      allOf:
        - if:
            properties:
              kind:
                const: extract_criteria
            required: [kind]
          then:
            properties:
              reportId:
                type: 'null'
              documentIds:
                minItems: 1
              projectDocumentIds:
                maxItems: 0
              previousReportIds:
                maxItems: 0
              criteriaSnapshotVersion:
                type: 'null'
        - if:
            properties:
              kind:
                const: analyze_report
            required: [kind]
          then:
            properties:
              reportId:
                type: string
                format: uuid
              documentIds:
                maxItems: 0
              criteriaSnapshotVersion:
                type: integer
                minimum: 1
              proposalCount:
                type: 'null'
        - if:
            properties:
              kind:
                const: extract_criteria
              status:
                const: succeeded
            required: [kind, status]
          then:
            properties:
              proposalCount:
                type: integer
                minimum: 0
    CriterionProposalReviewAction:
      type: string
      enum: [accept, correct, reject]
    CriterionProposalCorrection:
      type: object
      additionalProperties: false
      required: [code, description, sourceAnchors]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 4000
        deadline:
          type: [string, 'null']
          format: date
        sourceAnchors:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/SourceAnchor'
    CriterionProposalReview:
      type: object
      additionalProperties: false
      required: [proposalId, proposalRevision, action]
      properties:
        proposalId:
          type: string
          format: uuid
        proposalRevision:
          type: integer
          minimum: 1
        action:
          $ref: '#/components/schemas/CriterionProposalReviewAction'
        correction:
          $ref: '#/components/schemas/CriterionProposalCorrection'
        comment:
          type: string
          minLength: 1
          maxLength: 4000
      allOf:
        - if:
            properties:
              action:
                const: correct
            required: [action]
          then:
            required: [correction, comment]
          else:
            properties:
              correction: false
        - if:
            properties:
              action:
                const: reject
            required: [action]
          then:
            required: [comment]
    CriterionProposalReviewBatch:
      type: object
      additionalProperties: false
      required: [reviews]
      properties:
        reviews:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items:
            $ref: '#/components/schemas/CriterionProposalReview'
    CriterionProposalReviewRecord:
      type: object
      additionalProperties: false
      required:
        - id
        - proposalId
        - proposalRevision
        - action
        - correction
        - comment
        - createdCriterion
        - reviewedBy
        - reviewedAt
      properties:
        id:
          type: string
          format: uuid
        proposalId:
          type: string
          format: uuid
        proposalRevision:
          type: integer
          minimum: 1
        action:
          $ref: '#/components/schemas/CriterionProposalReviewAction'
        correction:
          oneOf:
            - $ref: '#/components/schemas/CriterionProposalCorrection'
            - type: 'null'
        comment:
          type: [string, 'null']
          maxLength: 4000
        createdCriterion:
          oneOf:
            - $ref: '#/components/schemas/Criterion'
            - type: 'null'
        reviewedBy:
          type: string
          minLength: 1
          maxLength: 255
        reviewedAt:
          type: string
          format: date-time
      allOf:
        - if:
            properties:
              action:
                enum: [accept, correct]
            required: [action]
          then:
            properties:
              createdCriterion:
                $ref: '#/components/schemas/Criterion'
        - if:
            properties:
              action:
                const: reject
            required: [action]
          then:
            properties:
              createdCriterion:
                type: 'null'
    CriterionProposal:
      type: object
      additionalProperties: false
      required:
        - id
        - analysisJobId
        - projectId
        - revision
        - proposedCode
        - proposedDescription
        - proposedDeadline
        - sourceAnchors
        - review
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        analysisJobId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        revision:
          type: integer
          minimum: 1
        proposedCode:
          type: string
          minLength: 1
          maxLength: 100
        proposedDescription:
          type: string
          minLength: 1
          maxLength: 4000
        proposedDeadline:
          type: [string, 'null']
          format: date
        sourceAnchors:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/SourceAnchor'
        review:
          oneOf:
            - $ref: '#/components/schemas/CriterionProposalReviewRecord'
            - type: 'null'
        createdAt:
          type: string
          format: date-time
    CriterionProposalReviewBatchResult:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/CriterionProposalReviewRecord'
...
    PaginatedCriterionProposals:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CriterionProposal'
        nextCursor:
          type: [string, 'null']
          minLength: 1
          maxLength: 2048
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 255
        code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 1000
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, requestId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          minLength: 1
          maxLength: 255
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
          maxLength: 2000
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          minLength: 1
          maxLength: 100
        requestId:
          type: string
          minLength: 1
          maxLength: 128
        errors:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/FieldError'